+++ tests/e2e/pubsub.subscription.spec.ts
@@ -11,9 +11,9 @@ async function createTopic(page: any, topicName: string) {
 
 test.describe('PubSub Subscription', () => {
   // Connect and navigate before each test
-  test.beforeEach(async ({ page, pubsubConnection, deletePubSubResources }) => {
+  test.beforeEach(async ({ page, emulatorConnection, deletePubSubResources }) => {
     await deletePubSubResources.delete();
-    await pubsubConnection.connect('/', 'localhost', '8085', 'project_test');
+    await emulatorConnection.connect('pubsub', 'localhost', '8085', 'project_test');
     await page.locator('#PubSub a').click();
     // Go to the Subscription page
     await page.locator('button[role="tab"]:has-text("Subscription")').click();