+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2025-xx-xx
+
+ - new setup page for all emulators type
+ - new home page 
+ - add BigQuery emulator (not from Google) and Firestore from Docker compose
+ - start implementatation for BigQuery & Firestore
+ - new rules for eslint to sort import
+ - Update all dependancies
+
 ## [0.5.1] - 2025-04-26
 
  - Improve app icon